Tests: Adjust according to "connectField" changes
